requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
qrcode
redis
